+++ library/core/src/main/java/com/google/android/exoplayer2/drm/ExoMediaDrm.java
@@ -33,9 +33,42 @@ import java.util.UUID;
  */
 public interface ExoMediaDrm<T extends ExoMediaCrypto> {
 
+  /** {@link ExoMediaDrm} instances provider. */
+  interface Provider<T extends ExoMediaCrypto> {
+
+    /**
+     * Returns an {@link ExoMediaDrm} instance with acquired ownership for the DRM scheme identified
+     * by the given UUID.
+     *
+     * <p>Each call to this method must have a corresponding call to {@link ExoMediaDrm#release()}
+     * to ensure correct resource management.
+     */
+    ExoMediaDrm<T> acquireExoMediaDrm(UUID uuid);
+  }
+
   /**
-   * @see MediaDrm#EVENT_KEY_REQUIRED
+   * {@link Provider} implementation which provides an {@link ExoMediaDrm} instance owned by the
+   * app.
+   *
+   * <p>This provider should be used to manually handle {@link ExoMediaDrm} resources.
    */
+  final class AppManagedProvider<T extends ExoMediaCrypto> implements Provider<T> {
+
+    private final ExoMediaDrm<T> exoMediaDrm;
+
+    /** Creates an instance, which provides the given {@link ExoMediaDrm}. */
+    public AppManagedProvider(ExoMediaDrm<T> exoMediaDrm) {
+      this.exoMediaDrm = exoMediaDrm;
+    }
+
+    @Override
+    public ExoMediaDrm<T> acquireExoMediaDrm(UUID uuid) {
+      exoMediaDrm.acquire();
+      return exoMediaDrm;
+    }
+  }
+
+  /** @see MediaDrm#EVENT_KEY_REQUIRED */
   @SuppressWarnings("InlinedApi")
   int EVENT_KEY_REQUIRED = MediaDrm.EVENT_KEY_REQUIRED;
   /**
@@ -235,6 +268,16 @@ public interface ExoMediaDrm<T extends ExoMediaCrypto> {
   Map<String, String> queryKeyStatus(byte[] sessionId);
 
   /**
+   * Acquires ownership over this instance, which must be released by calling {@link #release()}.
+   */
+  void acquire();
+
+  /**
+   * Releases ownership of this instance. If a call to this method causes this instance to have no
+   * acquired ownerships, releases the underlying resources.
+   *
+   * <p>Callers of this method must not make any further use of this instance.
+   *
    * @see MediaDrm#release()
    */
   void release();
@@ -272,6 +315,10 @@ public interface ExoMediaDrm<T extends ExoMediaCrypto> {
    */
   T createMediaCrypto(byte[] sessionId) throws MediaCryptoException;
 
-  /** Returns the {@link ExoMediaCrypto} type created by {@link #createMediaCrypto(byte[])}. */
+  /**
+   * Returns the {@link ExoMediaCrypto} type created by {@link #createMediaCrypto(byte[])}, or null
+   * if this instance cannot create any {@link ExoMediaCrypto} instances.
+   */
+  @Nullable
   Class<T> getExoMediaCryptoType();
 }